requirements.txt
accelerate==1.12.0
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1763409923949/work
bitsandbytes==0.49.1
brotli==1.2.0
cachetools==7.0.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
contourpy==1.3.2
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
debugpy @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_debugpy_1769744974/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
descartes==1.1.0
diffusers==0.36.0
distro==1.9.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1763918002538/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
fastapi==0.128.1
ffmpy==1.0.0
filelock==3.20.0
fire==0.7.1
fonttools==4.61.1
fsspec==2025.12.0
gradio==6.5.1
gradio_client==2.0.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.1
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipykernel_1761567932/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1748711175/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jupyter_client_1767954655/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jupyter_core_1760643864/work
kiwisolver==1.4.9
lazy_loader==0.4
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1761214490209/work
mdurl==0.1.2
mpmath==1.3.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.4.2
numpy==1.26.4
nuscenes-devkit==1.2.0
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
openai==2.16.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.11.7
packaging @ file:///home/task_176104877067765/conda-bld/packaging_1761049113113/work
pandas==2.3.3
parameterized==0.9.0
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1755974222/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==12.0.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1764950726/work
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
protobuf==3.20.0
psutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_psutil_1769678154/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pycocotools==2.0.11
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pyparsing==3.3.2
pyquaternion==0.9.9
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-multipart==0.0.22
pytorch-triton==3.2.0+git4b3bb1f8
pytz==2025.2
PyYAML==6.0.3
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pyzmq_1757387023/work
regex==2026.1.15
requests==2.32.5
rich==14.3.2
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
Shapely==1.8.5.post1
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
sniffio==1.3.1
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starlette==0.50.0
sympy==1.14.0
termcolor==3.3.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1765836373149/work
tqdm==4.67.3
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==5.0.0
triton==3.6.0
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1769858610910/work
zipp==3.23.0
